api/requirements.txt
alembic==1.6.4
attrs==21.2.0
cachelib==0.1.1
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.0.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
formelsammlung==5.0.0
greenlet==1.1.0
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
packaging==20.9
Pillow==9.0.1
pluggy==0.13.1
py==1.10.0
pydocstyle==5.1.1
PyMySQL==1.0.2
pyparsing==2.4.7
PyQRCode==1.2.1
pytest==6.2.4
pytest-flask==1.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==6.0
six==1.16.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.15
toml==0.10.2
Werkzeug==2.0.1
WTForms==2.3.3
